+++ examples/test-es6-proxy.js
@@ -1,33 +1,27 @@
+var foo = {}
+
+/**
+ * @type {Boolean}
+ */
+foo.bar	= true;
+
+foo._jsdocedProperties = foo._jsdocedProperties || {}
+foo._jsdocedProperties.bar = function(value){ return typeof value === 'boolean' }
+
 var Proxy = require('harmony-proxy');
 
-var validator = {
+foo = new Proxy(foo, {
 	set: function(object, property, value) {
-		if (property === 'age') {
-			if ( Number.isInteger(value) === false ) {
-				throw new TypeError('The age is not an integer');
-			}
-			if (value > 200) {
-				throw new RangeError('The age seems invalid');
-			}
+		if( object._jsdocedProperties && object._jsdocedProperties[property] ){
+			var checkingFunction = object._jsdocedProperties[property]
+			var isValid = checkingFunction(value)
+			console.assert('property', property, 'value', value)			
+			// console.log('checking property', property, 'isValid', isValid)
 		}
-		
 		// The default behavior to store the value
 		object[property] = value;
 	}
-};
+})
 
-
-var MyClass = function(){
-	
-}
-
-var myClassInstance = new MyClass()
-
-var person = new Proxy(myClassInstance, validator);
-
-console.log('bla', person instanceof MyClass)
-
-person.age = 100;
-console.log(person.age); // 100
-person.age = 'young'; // Throws an exception
-// person.age = 300; // Throws an exception
+foo.bar = false
+foo.bar = 100